Move all SideNavigation logic from Faq to SideNavigationOverview and use slot to display FaqCategory component
